refactor: Move `launch_northstar` to own module
Part of #329
